aihub_automate/requirements.txt
asgiref==3.4.1
asttokens==2.0.8
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
colorama==0.4.5
decorator==5.1.1
dj-database-url==0.5.0
Django==3.2.13
django-bootstrap-v5==1.0.11
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-environ==0.8.1
django-js-asset==2.0.0
django-phone-field==1.8.1
django-phonenumber-field==7.0.0
django-use-email-as-username==1.2.0
django-widget-tweaks==1.4.12
executing==1.0.0
gunicorn==20.1.0
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
phonenumbers==8.12.55
pickleshare==0.7.5
prompt-toolkit==3.0.31
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.13.0
PyPDF2==2.11.1
pytz==2021.3
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.0
toml==0.10.2
traitlets==5.4.0
tzdata==2022.2
wcwidth==0.2.5
whitenoise==5.3.0
